Fix crash when trying to build a road on a map with at least 65536 nodes, reported at [http://wl.widelands.org/forum/topic/57/].
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3968 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/pathfield.cc
@@ -25,7 +25,7 @@
 
 namespace Widelands {
 
-Pathfields::Pathfields(uint16_t const nrfields)
+Pathfields::Pathfields(uint32_t const nrfields)
 	: fields(new Pathfield[nrfields]), cycle(0)
 {}
